+++ test_regress/t/t_dotfiles.py
@@ -18,8 +18,7 @@ test.compile(v_flags2=["--dumpi-graph 6"], threads=2)
 
 for dotname in [
         "linkcells", "task_call", "gate_graph", "gate_final", "acyc_simp", "orderg_pre",
-        "orderg_acyc", "orderg_order", "orderg_domain", "ordermv_initial", "ordermv_hazards",
-        "ordermv_contraction", "ordermv_transitive1", "orderg_done", "pack", "schedule"
+        "orderg_acyc", "orderg_order", "orderg_domain", "orderg_done", "pack", "schedule"
 ]:
     # Some files with identical prefix are generated multiple times during
     # Verilation. Ensure that at least one of each dotname-prefixed file is generated.
